requirements.txt
absl-py==0.11.0
aiohttp==3.6.3
appdirs==1.4.4
argon2-cffi==20.1.0
astunparse==1.6.3
async-timeout==3.0.1
attrs==19.3.0
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==3.1.5
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.1
cfgv==3.1.0
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
discord.py==1.5.1
distlib==0.3.1
entrypoints==0.3
filelock==3.0.12
future==0.18.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
identify==1.4.21
idna==2.9
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
iso8601==0.1.13
jedi==0.17.1
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
Keras-Preprocessing==1.1.2
lxml==4.5.1
Markdown==3.3.3
MarkupSafe==1.1.1
mistune==0.8.4
multidict==4.7.6
nbconvert==5.6.1
nbformat==5.0.7
nodeenv==1.4.0
notebook==6.1.1
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
pprintpp==0.4.0
pre-commit==2.6.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio @ file:///D:/Downloads/PyAudio-0.2.11-cp38-cp38-win_amd64.whl
pycparser==2.20
Pygments==2.6.1
PyNaCl==1.3.0
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.1
pywin32==228
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.5.4
selenium==3.141.0
Send2Trash==1.5.0
serial==0.0.97
six==1.15.0
sounddevice==0.4.1
soupsieve==2.0.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.9
virtualenv==20.0.25
wcwidth==0.2.4
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.5.1
ytmusicapi==0.12.0
